update api update and delete campaign

+++ prisma/migrations/20250124113933_campaigns/migration.sql
@@ -4,7 +4,7 @@ CREATE TABLE "Campaign" (
     "name" TEXT NOT NULL,
     "description" TEXT NOT NULL,
     "image" TEXT NOT NULL,
-    "donation_target" BIGINT NOT NULL,
+    "donation_target" INT NOT NULL,
     "category" TEXT NOT NULL,
     "status" TEXT NOT NULL,
     "start_date" TIMESTAMP(3) NOT NULL,
